+++ src/main/java/com/apap/tugas01/repository/InstansiDB.java
@@ -1,6 +1,7 @@
 package com.apap.tugas01.repository;
 
 import java.math.BigInteger;
+import java.util.Set;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
@@ -8,6 +9,5 @@ import org.springframework.stereotype.Repository;
 import com.apap.tugas01.model.InstansiModel;
 
 @Repository
-public interface InstansiDB extends JpaRepository<InstansiModel, BigInteger>{
-
+public interface InstansiDB extends JpaRepository<InstansiModel, Long>{
 }
